server/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2-binary
Werkzeug
SQLAlchemy
alembic
gunicorn
huggingface_hub
Pillow
google-generativeai
requests
google-genai
